+++ src/simmate/apps/lab_automation/controllers/__init__.py
@@ -1,1 +1,6 @@
+# -*- coding: utf-8 -*-
+
+from .fuzzy_logic import FuzzyLogic
+from .on_off import OnOff
 from .pid import PID
+from .time_proportional import TimeProportional